+++ CHANGELOG.md
@@ -6,6 +6,22 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 - Feature: `Tokens`: Add `typography` css tokens.
 
+## [11.6.1] - 2022-07-04
+
+Fix:
+
+- `Autocomplete`: Update updateSuggestionsStrategy function.
+
+## [11.6.0] - 2022-07-01
+
+Feature: 
+
+- `Autocomplete`: Can pass objects to autocomplete items.
+
+Fix:
+
+- `Button`: New `hydrogen` modifier border style.
+
 ## [11.5.2] - 2022-06-28
 
 Fix: